+++ SpatialGDK/Source/SpatialGDK/Private/Utils/SpatialStatics.cpp
@@ -8,6 +8,7 @@
 #include "Interop/SpatialWorkerFlags.h"
 #include "Kismet/KismetSystemLibrary.h"
 #include "SpatialConstants.h"
+#include "EngineClasses/SpatialGameInstance.h"
 #include "SpatialGDKSettings.h"
 #include "Utils/InspectionColors.h"
 #include "Utils/SpatialActorGroupManager.h"
@@ -23,10 +24,10 @@ SpatialActorGroupManager* USpatialStatics::GetActorGroupManager(const UObject* W
 {
 	if (const UWorld* World = WorldContext->GetWorld())
 	{
-		if (const USpatialNetDriver* SpatialNetDriver = Cast<USpatialNetDriver>(World->GetNetDriver()))
+		if (const USpatialGameInstance* SpatialGameInstance = Cast<USpatialGameInstance>(World->GetGameInstance()))
 		{
-			check(SpatialNetDriver->ActorGroupManager.IsValid());
-			return SpatialNetDriver->ActorGroupManager.Get();
+			check(SpatialGameInstance->ActorGroupManager.IsValid());
+			return SpatialGameInstance->ActorGroupManager.Get();
 		}
 	}
 	return nullptr;
@@ -88,7 +89,13 @@ bool USpatialStatics::IsActorGroupOwnerForActor(const AActor* Actor)
 		return false;
 	}
 
-	return IsActorGroupOwnerForClass(Actor, Actor->GetClass());
+	const AActor* EffectiveActor = Actor;
+	while (EffectiveActor->bUseNetOwnerActorGroup && EffectiveActor->GetOwner() != nullptr)
+	{
+		EffectiveActor = EffectiveActor->GetOwner();
+	}
+
+	return IsActorGroupOwnerForClass(EffectiveActor, EffectiveActor->GetClass());
 }
 
 bool USpatialStatics::IsActorGroupOwnerForClass(const UObject* WorldContextObject, const TSubclassOf<AActor> ActorClass)
@@ -129,8 +136,13 @@ FName USpatialStatics::GetActorGroupForActor(const AActor* Actor)
 {
 	if (SpatialActorGroupManager* ActorGroupManager = GetActorGroupManager(Actor))
 	{
-		UClass* ActorClass = Actor->GetClass();
-		return ActorGroupManager->GetActorGroupForClass(ActorClass);
+		const AActor* EffectiveActor = Actor;
+		while (EffectiveActor->bUseNetOwnerActorGroup && EffectiveActor->GetOwner() != nullptr)
+		{
+			EffectiveActor = EffectiveActor->GetOwner();
+		}
+
+		return ActorGroupManager->GetActorGroupForClass(EffectiveActor->GetClass());
 	}
 
 	return SpatialConstants::DefaultActorGroup;