+++ CHANGELOG.md
@@ -65,6 +65,8 @@ Usage: `DeploymentLauncher createsim <project-name> <assembly-name> <target-depl
 - Added `bEnableClientQueriesOnServer` (defaulted false) which makes the same queries on the server as on clients if the unreal load balancer is enabled. Enable this to avoid clients seeing entities the server does not if the server's interest query has not been configured correctly.
 - The new Net Cull Distance client interest model is now enabled by default.
 - Added log warning when AddPendingRPC fails due to ControllerChannelNotListening.
+- When running with Offloading enabled, Actors will have local authority (ROLE_Authority) on servers for longer periods of time to allow more native Unreal functionality to work without problems.
+- When running with Offloading enabled, and trying to spawn Actors on a server which will not be the Actor Group owner for them, an error is logged and the Actor is deleted.
 
 ## Bug fixes:
 - Fixed a bug that caused the local API service to memory leak.